BUSINESS COMMUNICATION COURSE IEEE_REPORT
This research paper was made by 4 students including me on "OPTIMIZATION OF IoT AND AI TOOLS: REAL-TIME WASTE DETECTION AND IT’S IMPACT ON PUBLIC HEALTH".                                                                                                                                                    To respect the privacy of my groupmates and our course instructor, their names and IDs are not displayed in this repository. Only my name, student ID, and my contributions are shown.